Added bindings for types: landmark_map, bounding_box<int>, match_set
