+++ python/kwiver/vital/types/bounding_box.h
@@ -28,6 +28,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#ifndef PY_BBOX_H
+#define PY_BBOX_H
+
+#include <string.h>
 #include <vital/types/bounding_box.h>
 
 #include <Eigen/Core>
@@ -39,9 +43,8 @@
 namespace py = pybind11;
 
 
-typedef kwiver::vital::bounding_box<double> bbox;
-
-PYBIND11_MODULE(bounding_box, m)
+template<typename T>
+void bounding_box(py::module &m, const char * typestr)
 {
 
   /*
@@ -54,8 +57,12 @@ PYBIND11_MODULE(bounding_box, m)
 
    *
    */
+  typedef kwiver::vital::bounding_box<T> bbox;
+  char pyclass_name[20];
+  strcpy(pyclass_name, "BoundingBox");
+  strcat(pyclass_name, typestr);
 
-  py::class_<bbox, std::shared_ptr<bbox>>(m, "BoundingBox", R"(
+  py::class_<bbox, std::shared_ptr<bbox>>(m, pyclass_name, R"(
     Coordinate aligned bounding box.
 
     Example:
@@ -67,9 +74,9 @@ PYBIND11_MODULE(bounding_box, m)
         4000.0
 
     )")
-  .def(py::init<Eigen::Matrix<double,2,1>, Eigen::Matrix<double,2,1>>())
-  .def(py::init<Eigen::Matrix<double,2,1>, double, double>())
-  .def(py::init<double, double, double, double>(), py::doc(R"(
+  .def(py::init<Eigen::Matrix<T,2,1>, Eigen::Matrix<T,2,1>>())
+  .def(py::init<Eigen::Matrix<T,2,1>, T, T>())
+  .def(py::init<T, T, T, T>(), py::doc(R"(
         Create a box from four coordinates
 
         Args:
@@ -119,3 +126,6 @@ PYBIND11_MODULE(bounding_box, m)
     })
   ;
 }
+
+
+#endif
