+++ python/kwiver/vital/types/match_set.cxx
@@ -28,30 +28,40 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * \file algorithm_implementation.cxx
- *
- * \brief python bindings for algorithm
- */
-
 #include <pybind11/pybind11.h>
-#include <python/kwiver/vital/types/bounding_box.h>
-#include <python/kwiver/vital/types/image.h>
-#include <python/kwiver/vital/types/image_container.h>
-#include <python/kwiver/vital/types/landmark.h>
-#include <python/kwiver/vital/types/landmark_map.h>
+#include <vital/types/match_set.h>
+#include <pybind11/stl.h>
+#include <pybind11/stl_bind.h>
+#include <pybind11/embed.h>
 #include <python/kwiver/vital/types/match_set.h>
 
-namespace py = pybind11;
+void match_set(py::module &m) {
+    py::bind_vector<std::vector<match_t>>(m, "MatchVector");
+    py::class_<match_set_t, std::shared_ptr<match_set_t>>(m, "BaseMatchSet")
+    .def("size", &match_set_t::size)
+    .def("matches", &match_set_t::matches, py::return_value_policy::reference_internal)
 
-PYBIND11_MODULE(types, m)
-{
-  bounding_box<int>(m, "Int");
-  bounding_box<float>(m, "Float");
-  bounding_box<double>(m, "Double");
-  image(m);
-  image_container(m);
-  landmark(m);
-  landmark_map(m);
-  match_set(m);
+    .def("__repr__", [](py::object& self) -> std::string {
+        auto locals = py::dict(py::arg("self")=self);
+        py::exec(R"(
+            classname = self.__class__.__name__
+            retval = '<%s at %s>' % (classname, hex(id(self)))
+            )", py::globals(), locals);
+        return locals["retval"].cast<std::string>();
+    })
+
+    .def("__str__", [](py::object& self) -> std::string {
+        auto locals = py::dict(py::arg("self")=self);
+        py::exec(R"(
+            classname = self.__class__.__name__
+            retval = '<%s>' % (classname)
+            )", py::globals(), locals);
+        return locals["retval"].cast<std::string>();
+    });
+
+
+    py::class_<s_match_set_t, match_set_t, std::shared_ptr<s_match_set_t>>(m, "MatchSet")
+    .def(py::init<>())
+    .def(py::init<const std::vector<match_t>& >());
+
 }
